requirements.txt
asgiref==3.8.1
Django==5.1.2
django-filter==1.1.0
djangorestframework==3.15.2
djangorestframework-filters==0.11.1
shell==1.0.1
sqlparse==0.5.1
tzdata==2024.2
